pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "post-training-toolkit"
version = "0.2.0"
description = "Production-grade diagnostics & interpretability for RLHF training with Hugging Face TRL"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Post-Training Toolkit Contributors" }
]
keywords = [
    "rlhf",
    "dpo",
    "ppo",
    "grpo",
    "sft",
    "llm",
    "transformers",
    "trl",
    "huggingface",
    "diagnostics",
    "machine-learning",
    "deep-learning",
    "post-training",
    "alignment",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "jinja2>=3.1.2",
    "scipy>=1.10.0",
    "python-dateutil>=2.8.2",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
trl = [
    "transformers>=4.35.0",
    "trl>=0.7.0",
    "torch>=2.0.0",
]
wandb = [
    "wandb>=0.15.0",
]
mlflow = [
    "mlflow>=2.0.0",
]
tensorboard = [
    "tensorboard>=2.14.0",
]
all = [
    "transformers>=4.35.0",
    "trl>=0.7.0",
    "torch>=2.0.0",
    "wandb>=0.15.0",
    "mlflow>=2.0.0",
    "tensorboard>=2.14.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
ptt-diagnose = "post_training_toolkit.cli:diagnose_cli"
ptt-compare = "post_training_toolkit.cli:compare_cli"
ptt-validate-resume = "post_training_toolkit.cli:validate_resume_cli"
ptt-agent-diagnose = "post_training_toolkit.cli:agent_diagnose_cli"

[project.urls]
Homepage = "https://github.com/your-org/post-training-toolkit"
Documentation = "https://github.com/your-org/post-training-toolkit#readme"
Repository = "https://github.com/your-org/post-training-toolkit"
Issues = "https://github.com/your-org/post-training-toolkit/issues"
Changelog = "https://github.com/your-org/post-training-toolkit/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.sdist]
include = [
    "/post_training_toolkit",
    "/templates",
]

[tool.hatch.build.targets.wheel]
packages = ["post_training_toolkit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "-ra",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["post_training_toolkit"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hatch
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG", "S101"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
